requirements.txt
streamlit==1.37.1
pandas==2.2.2
pillow==10.4.0
pytesseract==0.3.10
opencv-python-headless==4.10.0.84
